Fully functional Profile-edit feature
Users can now edit their profile using the profile/edit route
which maps to the profile-edit component. I think this might
be overkill, as I think I could've just done it in the profile/edit
route without the component, but I might re-use the profile-edit
component and rename it in the future.