BPA! Grab all profiles when I get all games
In the /games route I want to display all the users'
names and nationalities alongside them. In order to
do that I need to load all Profiles on load of /games.
BEST PRACTICE ALERT! I'm not sure if this is the best
way to do it, but it's working now, so I'm going with it
until I learn a better way.