Update sign-up to sign-in twice now
Sign up used to sign-up, sign-in, create profile.
However, this prevented the profile from being stored locally
in the profile_id section of the auth service. I think what I actually
need to do is set auth.credentials.profile_id after I create and save
the new profile that was created, but I know this will work for now.

Now after profile create it signs in again. So the new process is
1. Sign Up
2. Sign In
3. Create Profile
4. Sign in again to update the profile id in `auth`