Adds anAuth option
For games that are trying to be joined by someone that isn't logged
in (doesn't have a `profile_id` defined), their action on `join-game`
is diverted from `join-game` to `sign-in` so they can sign in to
perform this action. Future feature might be sending them back to the
page they came from after they've logged in. Something in the store
about last-visited??