Allow joining and leaving games from either view
Game and Games views were acting differently. They're still
acting differently in that when I join a game in the `game` view
I have to "force refresh" the game by transitioning to games and
then back to game. It's definitely not the best way to do it, but
it works currently. Next step is to try to bind the game AND the
user down to the game route so I can join them up when it's time.